// src/app/api/ui/enums/justification-for-non-use-national-mode-type.enum.ts
export enum JustificationForNonUseNationalModeTypeEnum {
  None = 0,
  Type5a = 1,
  Type5v = 2,
  Type5g = 3,
  Type5d = 4,
  Type5e = 5,
  Type5j = 6,
  Type5z = 7,
  Type5i = 8,
  Type5k = 9,
  Type5l = 10,
  Type5m = 11,
  Type5n = 12,
  Type6a = 13,
  Type6b = 14,
  Type6v = 15,
  Type6g = 16,
  Other = 17,
}